Add the es-419 language code to support South American Spanish
This code is useful for targeting Spanish spoken in the Latin America
and the Caribbean region. There are no plans to make this available as
an interface language, hence I am not adding a language file with a
fallback to 'es'.

Bug: T112889
Cherry-picked from: a19320bf90b58437020a4bd50803e0d29a8a9269

Change-Id: Ib7024fda56799d0a0632761aff91574fff6a1dee