maintenance: Update update-oojs.sh to use misc files from package
These are already included in the package. Using them from the dist/
directory is somewhat against the way things are normally done in NPM,
given the repository itself already becomes the package.

If we were building a separate directory for a tarball, that
would make more sense, but that isn't the case.

Note: In verions prior to OOjs v2.2.1, all files were included in the package,
version v2.2.1 added the 'files' whitelist which included only dist/, LICENSE*,
README*, History* and package.json, which in particular means AUTHORS.txt
is missing. This was fixed in v2.2.2 on the same day. MediaWiki will upgrade
from v2.2.0 to v2.2.2 directly.

Change-Id: I4bd91811c754addac6c9826e91580eff47c7c5d6