services: Add linebreak between regular and spread parameter
Follows-up 1e680456b4, which removed some array_merge/call_user_func
abstraction, but improved readability a lot, but the absence of a
line break makes it slightly less obvious how and what parameters
are being passed. It's usually either comma-separated segments on
the same line that as the open parenthesis, or one expression per
line.

Change-Id: Ic388cf41ed3ed60d9b47c04392242e40cdcb69b0