Introduce constants for fetchLanguageName(s)
Before:
 Language::fetchLanguageNames( null, 'mwfile' );
After:
 Language::fetchLanguageNames( self::AS_AUTONYMS, self::SUPPORTED );

The latter is much more readable and doesn't require knowing the
method definition by heart when seeing it in the code.

I did not add a constant for 'mw' because:
* It is the default value for a last parameter, so it can be left out
* In the long run I expect it will be folded into 'all', see T190129

I did not update any callers; that can be done in follow-ups after
the naming is agreed upon.

Change-Id: Ia9bbb525f213046487f08bcd196876befe87cae8