Checkboxes should be styled on non-JS browsers
Previously we didn't do this so that checkboxes worked on older
versions of Opera Mini.

This was added 3 years ago
(https://trello.com/c/D2FnYlmf/21-bug-fix-broken-settings-page-on-opera-mini)
https://imgur.com/a/DL6ASy4

This no longer seems to be a problem on more modern versions of
Opera Mini and given how this degrades the experience of more modern
browsers I propose we restore this styling.

Bug: T196947
Change-Id: I5f646b5c939e498eda9bec249658a969f3f91c5d