mediawiki.inspect: Use property assignments for clarity
This changes the assignment of these methods from being declared
at once in an object literal, to assigning properties in separate
statements.

This is functionally a no-op and produces the exact same values,
but was factored out of the next commit to keep the diff cleaner
when fixing T197810 (which changes the host object from a plain
empty object, to a function object, which cannot be declared
as a literal).

Bug: T197810
Change-Id: I6ba1891acf8807cb9e01fa11a608e0c509e5f2bc