RC Filters: Check if app is initialized before calling change tags
This only occurs when highlighting is enabled in the query parameters on page
load. If you load a watchlist page with highlighting enabled, you'll see that
FilterTagMultiselectWidget calls the parent onChangeTags method, and then we
end up at updateIfHeightChanged where a resize event occurs. This commit
checks to see if the app is still initializing before propagating this event.

Bug: T198399
Change-Id: I8d36875ecf859b40854fe57fecf824e648bb3bf8