RCFilters: Hide highlight containers when RCFilters is disabled
The mw-rcfilters-ui-highlights divs are given display: none; (if
highlights are disabled) or display: inline-block; (if they are enabled)
by the RCFilters CSS, but none of that CSS loads when RCFilters is
disabled, so in that case they default to display: block;. I don't know
why this doesn't cause them to break the line in Chrome and Firefox, but
I suspect that this is what causes the display issues in Edge that
people have been reporting.

This is a speculative fix because I'm not able to run Edge to verify.

Bug: T198440
Change-Id: I7b4d9b01d9b78828c273895a0093a2a7c0b4fc99