build: Set Travis conf to "allow failure" for long-failing builds
Specifically, PHP 7.2, postgres and all HHVM builds have been
failing for at least a month or more. In order for notifications
to become useful, whitelist these for now, so that next time
the main MySQL/PHP 7.0 or PHP 7.1 build failes, we'll notice
sooner.

Also move the notifs back to #wikimedia-dev and make it match
the configuration of other repos that use Travis via GitHub
(such as CSSJanus), specifically, to use a message template
that spans only a single line, instead of the default which
spans several lines.

Change-Id: I3abbff51414d33cde34cca26896724e8999c32db