Avoid duplicate IDs on Special:Preferences
The value of the 'section' parameter is used to generate some ID
attributes on Special:Preferences. Awkwardly, only the second part
(after the slash) is used, so have to ensure these are unique.

It is also used for the default localisation message for the title of
the section. Override this so that we don't have to rename and
duplicate the messages.

Bug: T198875
Change-Id: I0f22aafc8b2ad860dd42c5373eafdf88e848e3ad