Remove hasOwnProperty checks in for-loops
There are some good uses of hasOwnProperty checks, but in for-loops
they only make sense if the object in question is not a plain
object (eg. class instance) and the code in question is explicitly
interested in instance properties. This is rare, but one example
exists in mediawiki.Uri, where it clones an instance by copying
properties from another instance.

Anywhere a plain object is looped over, the check is redundant
because Object.prototype has no enumerable own properties.

Both jQuery and ResourceLoader do not support environments
that alter Object.prototype to be that way, and the majority
of our for-loops actually already didn't have this check.

This commit removes the few hasOwnProperty checks that still
existed on for-loops over plain objects.

Side-note: jQuery.each() does an unfiltered for-loop, which means
           it too does not perform hasOwnProperty checks.

Change-Id: Ib5e5a04a5a8e21ab62b4e779117077ea5b0f8a47