installer: Suggest using 127.0.0.1 in the db not found error message
Its a really common error people have where they specify
localhost, not realizing that mysql does socket magic for that
value, and they need to specify 127.0.0.1 instead.

Also remove windows specific language in the help text.
Although more of an issue on windows, it is not unheard of
to happen on linux if the mysql socket is not configured.

Change-Id: Ie8199ab50d37dc2bc43dfafe2e5ef7812e9d832c