Clarify warnings about missing software diff3 and git.
Sometimes we get people on irc freaking out about the warnings.
If diff3 is missing, you just get a lot of edit conflicts.

If git is missing (A somewhat more common case) it just means that
Special:Version won't show the commit hash for things. Which
first of all doesn't matter and secondly is irrelavent to 90% of
users who are not installing from git.

Change-Id: I9c9ea797a80bd4efb6826f94373deee0311dc83a