mediawiki.inspect: Add headings to console and use console.group()
This adds a log message before each dump to indicate which report follows.
Most of the table's headings communicated to some extend what the table is
about, but the heading should make it easier to use.

Also, where available, use console.group(), which makes the
tables collapsible for easier navigation.

Change-Id: I6855441ce6e3052cdc875e439380820b61d040c1