Hard deprecate support for Parsoid v1 API
This was soft deprecated by f5c874cbaf2c1408342305fbc87dc6db36728e07 in
May 2015, and support for the v1 API was removed from Parsoid in May 2016,
but the compatibility code has persisted zombie-like in core.

Hard deprecate this now, to warn anyone who might still be using ancient
copies of Visual Editor and Parsoid, and then we can slice the code out
in the next MW release.

Bug: T100681
Change-Id: I68e0ab5b38f31b7708ba6f075dcb3328dde118d1