pager: Improve the extra sort field to work well when indexField is a string
The getExtraSortFields() function was well documented for the case when
getIndexField returns an array. But what if it returns a string?
Currently, using getExtraSortFields with a string as indexField means
that whatever you'll sort the table by, the extra sort will be used.
Instead, with this patch the extra sort will have two different
behaviours, depending on whether the extraSort array is associative or
not. If it is, then only the extra sort for the current field will be used.
Otherwise, like it already does, the extra sort will always be applied.

This is intended to be used by AbuseFilter in I979849e66 for T191694.

Change-Id: I0e695f96f18c7a9229753b1225dd473feb936a31