resourceloader: Remove support for `state(name, state)` signature
This has no usage in non-test code anywhere in Wikimedia Git,
and was only used in the test suite for mw.loader in core,
and in the test suite for the NavigationTiming extension.

The core usage is as part of this commit. The usage in NavTiming
is updated by I240ced4e65988f9.

Bug: T127328
Depends-On: I240ced4e65988f96c7ece3772378c2c9a335fb9a
Change-Id: Ic17c797e528feaf07a4777709d705615fea353e5